requirements.txt
-i https://pypi.org/simple
aiosmtplib==2.0.2; python_version >= '3.7' and python_version < '4.0'
aiosqlite==0.17.0; python_version >= '3.6'
annotated-types==0.6.0; python_version >= '3.8'
anyio==4.3.0; python_version >= '3.8'
async-timeout==4.0.3; python_full_version < '3.12.0'
asyncpg==0.29.0; python_full_version >= '3.8.0'
bcrypt==4.1.2; python_version >= '3.7'
blinker==1.7.0; python_version >= '3.8'
boto3==1.34.65; python_version >= '3.8'
botocore==1.34.65; python_version >= '3.8'
certifi==2024.2.2; python_version >= '3.6'
chardet==5.2.0; python_version >= '3.7'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
dnspython==2.6.1; python_version >= '3.8'
email-validator==2.1.1; python_version >= '3.8'
exceptiongroup==1.2.0; python_version >= '3.7'
fastapi==0.110.0; python_version >= '3.8'
fastapi-mail==1.4.1; python_version < '4.0' and python_full_version >= '3.8.1'
h11==0.14.0; python_version >= '3.7'
httptools==0.6.1; python_full_version >= '3.8.0'
idna==3.6; python_version >= '3.5'
iso8601==1.1.0; python_version < '4.0' and python_full_version >= '3.6.2'
jinja2==3.1.3; python_version >= '3.7'
jmespath==1.0.1; python_version >= '3.7'
mangum==0.17.0; python_version >= '3.7'
markupsafe==2.1.5; python_version >= '3.7'
numpy==1.26.4; python_version < '3.11'
pandas==2.2.2; python_version >= '3.9'
passlib==1.7.4
pydantic==2.6.4; python_version >= '3.8'
pydantic-core==2.16.3; python_version >= '3.8'
pydantic-settings==2.2.1; python_version >= '3.8'
pyjwt==2.4.0; python_version >= '3.6'
pypika-tortoise==0.1.6; python_version >= '3.7' and python_version < '4.0'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.1; python_version >= '3.8'
python-multipart==0.0.9; python_version >= '3.8'
pytz==2024.1
pyyaml==6.0.1; python_version >= '3.6'
requests==2.31.0; python_version >= '3.7'
s3transfer==0.10.1; python_version >= '3.8'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.1; python_version >= '3.7'
starlette==0.36.3; python_version >= '3.8'
tortoise-orm==0.20.0; python_version >= '3.8' and python_version < '4.0'
typing-extensions==4.10.0; python_version >= '3.8'
tzdata==2024.1; python_version >= '2'
urllib3==2.2.1; python_version >= '3.8'
uvicorn==0.28.0; python_version >= '3.8'
uvloop==0.19.0; python_full_version >= '3.8.0'
watchfiles==0.21.0; python_version >= '3.8'
websockets==12.0; python_version >= '3.8'
